+++ Auth.FWT.Data/Identity/IdentityFactory.cs
@@ -6,9 +6,9 @@ namespace Auth.FWT.Data.Identity
 {
     public class IdentityFactory
     {
-        public static RoleManager<UserRole, byte> CreateRoleManager(IEntitiesContext context)
+        public static RoleManager<UserRole, int> CreateRoleManager(IEntitiesContext context)
         {
-            var manager = new RoleManager<UserRole, byte>(new ApplicationRoleStore(context));
+            var manager = new RoleManager<UserRole, int>(new ApplicationRoleStore(context));
             return manager;
         }
